batchhandler: extract routing / setup logic
Handler funcs are now in their own file for easier organization
